refactored per initial peer review comments
