+++ CMakeLists.txt
@@ -6,4 +6,4 @@ set(CMAKE_BUILD_TYPE Release)
 
 include_directories(src ${Boost_INCLUDE_DIRS})
 
-add_executable(gfp_test src/gfp_test.cpp src/gfp.cpp)
+add_executable(gfp_test src/gfp_test.cpp src/gfp_generic.cpp)
